Move `Crypto` class and `StandardCrypto` to protocol-tpraos
Since #4788 `Crypto` and `StandardCrypto` only define `KES` and `VRF`,
which are primitives unused by ledger. For this reason they are now
being moved to `cardano-protocol-tpraos` package
