+++ eras/shelley-ma/test-suite/cardano-ledger-shelley-ma-test.cabal
@@ -107,7 +107,7 @@ test-suite cardano-ledger-shelley-ma-test
     cardano-ledger-shelley,
     cardano-ledger-shelley-ma-test,
     cardano-ledger-shelley-test,
-    cardano-protocol-tpraos >=1.0,
+    cardano-protocol-tpraos >=1.4,
     cardano-slotting,
     cardano-strict-containers,
     cborg,
