+++ eras/shelley-ma/test-suite/src/Test/Cardano/Ledger/EraBuffet.hs
@@ -1,5 +1,4 @@
 module Test.Cardano.Ledger.EraBuffet (
-  StandardCrypto,
   ShelleyEra,
   MaryEra,
   AllegraEra,
@@ -13,6 +12,5 @@ where
 
 import Cardano.Ledger.Allegra (AllegraEra)
 import Cardano.Ledger.Core (Era, Script, TxAuxData, TxBody, Value)
-import Cardano.Ledger.Crypto (StandardCrypto)
 import Cardano.Ledger.Mary (MaryEra)
 import Cardano.Ledger.Shelley (ShelleyEra)
