+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/ConcreteCryptoTypes.hs
@@ -12,8 +12,8 @@ module Test.Cardano.Ledger.Shelley.ConcreteCryptoTypes (
 where
 
 import Cardano.Crypto.KES (MockKES)
-import Cardano.Ledger.Crypto
 import Cardano.Ledger.Shelley (ShelleyEra)
+import Cardano.Protocol.Crypto
 import Cardano.Protocol.TPraos.API (PraosCrypto)
 import Test.Cardano.Protocol.Crypto.VRF.Fake (FakeVRF)
 
@@ -21,7 +21,7 @@ type C = ShelleyEra
 
 data MockCrypto
 
-instance Cardano.Ledger.Crypto.Crypto MockCrypto where
+instance Crypto MockCrypto where
   type KES MockCrypto = MockKES 10
   type VRF MockCrypto = FakeVRF
