+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Generator/Block.hs
@@ -18,11 +18,11 @@ where
 import qualified Cardano.Crypto.VRF as VRF
 import Cardano.Ledger.BHeaderView (bhviewBSize, bhviewHSize)
 import Cardano.Ledger.BaseTypes (UnitInterval)
-import Cardano.Ledger.Crypto (VRF)
 import Cardano.Ledger.Shelley.API
 import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.LedgerState (curPParamsEpochStateL)
 import Cardano.Ledger.Slot (SlotNo (..))
+import Cardano.Protocol.Crypto (VRF)
 import Cardano.Protocol.TPraos.API
 import Cardano.Protocol.TPraos.BHeader (
   BHeader (..),
