+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/UnitTests.hs
@@ -17,17 +17,8 @@ import qualified Cardano.Crypto.VRF as VRF
 import Cardano.Ledger.Address (Addr (..), raCredential, pattern RewardAccount)
 import Cardano.Ledger.BaseTypes hiding ((==>))
 import Cardano.Ledger.Coin
-import Cardano.Ledger.Credential (
-  Credential (..),
-  StakeReference (..),
- )
-import Cardano.Ledger.Crypto (StandardCrypto, VRF)
-import Cardano.Ledger.Keys (
-  KeyRole (..),
-  asWitness,
-  hashKey,
-  hashVerKeyVRF,
- )
+import Cardano.Ledger.Credential (Credential (..), StakeReference (..))
+import Cardano.Ledger.Keys (KeyRole (..), asWitness, hashKey)
 import Cardano.Ledger.PoolParams (
   PoolMetadata (..),
   PoolParams (..),
@@ -78,6 +69,7 @@ import Cardano.Ledger.Shelley.TxWits (
 import Cardano.Ledger.TxIn (TxIn (..))
 import qualified Cardano.Ledger.UMap as UM
 import Cardano.Ledger.Val ((<+>), (<->))
+import Cardano.Protocol.Crypto (StandardCrypto, VRF, hashVerKeyVRF)
 import Cardano.Protocol.TPraos.BHeader (checkLeaderValue)
 import Control.DeepSeq (rnf)
 import Control.State.Transition.Extended (PredicateFailure, TRC (..))
@@ -105,7 +97,6 @@ import Test.Cardano.Ledger.Core.KeyPair (
 import Test.Cardano.Ledger.Shelley.Arbitrary (
   ASC (ASC),
   StakeProportion (StakeProportion),
-  VRFNatVal (VRFNatVal),
  )
 import Test.Cardano.Ledger.Shelley.ConcreteCryptoTypes (C, MockCrypto)
 import Test.Cardano.Ledger.Shelley.Fees (sizeTests)
@@ -113,6 +104,7 @@ import Test.Cardano.Ledger.Shelley.Generator.Core (VRFKeyPair (..), genesisCoins
 import Test.Cardano.Ledger.Shelley.Generator.EraGen (genesisId)
 import Test.Cardano.Ledger.Shelley.Generator.ShelleyEraGen ()
 import Test.Cardano.Ledger.Shelley.Utils
+import Test.Cardano.Protocol.TPraos.Arbitrary (VRFNatVal (VRFNatVal))
 import Test.Control.State.Transition.Trace (checkTrace, (.-), (.->>))
 import qualified Test.QuickCheck.Gen as QC
 import qualified Test.QuickCheck.Random as QC
