+++ libs/cardano-protocol-tpraos/src/Cardano/Protocol/TPraos/Rules/OCert.hs
@@ -13,9 +13,10 @@ module Cardano.Protocol.TPraos.Rules.OCert (
 )
 where
 
+import Cardano.Crypto.KES (verifySignedKES)
 import Cardano.Ledger.BaseTypes
-import Cardano.Ledger.Crypto
 import Cardano.Ledger.Keys
+import Cardano.Protocol.Crypto
 import Cardano.Protocol.TPraos.BHeader
 import Cardano.Protocol.TPraos.OCert
 import Control.Monad.Trans.Reader (asks)
