+++ libs/cardano-protocol-tpraos/testlib/Test/Cardano/Protocol/Crypto/VRF.hs
@@ -8,11 +8,7 @@ module Test.Cardano.Protocol.Crypto.VRF (
 ) where
 
 import qualified Cardano.Crypto.VRF.Class as VRF
-import Cardano.Ledger.Crypto
-import Cardano.Ledger.Keys (
-  SignKeyVRF,
-  VerKeyVRF,
- )
+import Cardano.Protocol.Crypto
 
 data VRFKeyPair c = VRFKeyPair
   { vrfSignKey :: !(SignKeyVRF c)
